+++ app/Filament/Resources/Tags/Pages/EditTag.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace App\Filament\Resources\Tags\Pages;
+
+use Filament\Actions\ViewAction;
+use Filament\Actions\DeleteAction;
+use App\Filament\Resources\Tags\TagResource;
+use Filament\Pages\Actions;
+use Filament\Resources\Pages\EditRecord;
+
+class EditTag extends EditRecord
+{
+    protected static string $resource = TagResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            ViewAction::make(),
+            DeleteAction::make(),
+        ];
+    }
+}